refactor: move shop package from `no-meteor` to `node-add`
