+++ imports/node-app/core-services/shop/mutations/index.js
@@ -1,0 +1,5 @@
+import updateShop from "./updateShop.js";
+
+export default {
+  updateShop
+};
