+++ imports/node-app/core-services/shop/queries/test.js
@@ -1,4 +1,4 @@
-import queries from "./";
+import queries from "./index.js";
 
 const context = {
   dataLoaders: {
