+++ imports/node-app/core-services/shop/resolvers/Mutation/index.js
@@ -1,0 +1,5 @@
+import updateShop from "./updateShop.js";
+
+export default {
+  updateShop
+};
