+++ imports/node-app/core-services/shop/resolvers/Query/index.js
@@ -1,0 +1,11 @@
+import primaryShopId from "./primaryShopId.js";
+import primaryShop from "./primaryShop.js";
+import shop from "./shop.js";
+import shopBySlug from "./shopBySlug.js";
+
+export default {
+  primaryShopId,
+  primaryShop,
+  shop,
+  shopBySlug
+};
