+++ imports/node-app/core-services/shop/resolvers/Shop/index.js
@@ -1,9 +1,9 @@
 import { getXformedCurrencyByCode, getXformedCurrenciesByShop } from "@reactioncommerce/reaction-graphql-xforms/currency";
 import { encodeNavigationTreeOpaqueId } from "@reactioncommerce/reaction-graphql-xforms/navigationTree";
 import { encodeShopOpaqueId } from "@reactioncommerce/reaction-graphql-xforms/shop";
-import brandAssets from "./brandAssets";
-import defaultNavigationTree from "./defaultNavigationTree";
-import tags from "./tags";
+import brandAssets from "./brandAssets.js";
+import defaultNavigationTree from "./defaultNavigationTree.js";
+import tags from "./tags.js";
 
 export default {
   _id: (node) => encodeShopOpaqueId(node._id),
