+++ imports/node-app/core-services/shop/resolvers/index.js
@@ -1,0 +1,9 @@
+import Mutation from "./Mutation/index.js";
+import Query from "./Query/index.js";
+import Shop from "./Shop/index.js";
+
+export default {
+  Mutation,
+  Query,
+  Shop
+};
