--- imports/plugins/core/shop/server/mutations/index.js
@@ -1,5 +1,0 @@
-import updateShop from "./updateShop";
-
-export default {
-  updateShop
-};
