--- imports/plugins/core/shop/server/resolvers/Mutation/index.js
@@ -1,5 +1,0 @@
-import updateShop from "./updateShop";
-
-export default {
-  updateShop
-};
