--- imports/plugins/core/shop/server/resolvers/Query/index.js
@@ -1,11 +1,0 @@
-import primaryShopId from "./primaryShopId";
-import primaryShop from "./primaryShop";
-import shop from "./shop";
-import shopBySlug from "./shopBySlug";
-
-export default {
-  primaryShopId,
-  primaryShop,
-  shop,
-  shopBySlug
-};
