--- imports/plugins/core/shop/server/resolvers/index.js
@@ -1,9 +1,0 @@
-import Mutation from "./Mutation";
-import Query from "./Query";
-import Shop from "./Shop";
-
-export default {
-  Mutation,
-  Query,
-  Shop
-};
